% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamy-package.R
\docType{package}
\name{streamy-package}
\alias{streamy}
\alias{streamy-package}
\title{streamy: Inline Asynchronous Generator Results into Documents}
\description{
Given a 'coro' asynchronous generator instance that produces text or 'ellmer' content, write its text into a document selection in 'RStudio' and 'Positron'. This is particularly helpful for streaming large language model responses into the user's editor.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/streamy}
  \item \url{https://simonpcouch.github.io/streamy/}
  \item Report bugs at \url{https://github.com/simonpcouch/streamy/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
