% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{stream}
\alias{stream}
\title{Stream generator results into a document}
\usage{
stream(
  generator,
  context = rstudioapi::getActiveDocumentContext(),
  interface = c("prefix", "replace", "suffix")
)
}
\arguments{
\item{generator}{A \code{\link[coro:generator]{coro::generator()}} function.}

\item{context}{Optional. An RStudio document context.}

\item{interface}{One of \code{"prefix"}, \code{"replace"}, or \code{"suffix"}, describing
how to the active selection will be interfaced with. Defaults to \code{"replace"}.}
}
\value{
Nothing; called for its side effect, modifying the context of the current
selection with results from the generator.
}
\description{
Stream generator results into a document
}
