% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamy-package.R
\docType{package}
\name{streamy-package}
\alias{streamy}
\alias{streamy-package}
\title{streamy: Inline Contents from Generators into Documents}
\description{
Given a 'coro' generator, inlines results into a document selection in 'RStudio' and 'Positron'. This is particularly helpful for streaming LLM responses into documents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/streamy}
  \item \url{https://simonpcouch.github.io/streamy/}
  \item Report bugs at \url{https://github.com/simonpcouch/streamy/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

}
\keyword{internal}
